+++ backend/src/controllers/transaction_controllers/shopController.ts
@@ -1,17 +1,14 @@
 import { Request, Response } from 'express';
-import { PaymentRequest } from '../../service/Xendit Service/xenditClient';
+import { PaymentRequest } from '../../service/xenditClient';
 import { PaymentRequestParameters, PaymentRequestCurrency } from 'xendit-node/payment_request/models';
 import User from '../../model/user'; // Import the User model
-import Item from '../../model/ItemModel'; // Import the Item model
+import Item from '../../model/itemModel'; // Import the Item model
 import dotenv from 'dotenv';
-import { createPaymentMethod } from '../../service/Xendit Service/Subscription/paymentMethodService';
-import { getCoinBalance, deductCoinsByTokens } from '../../service/Xendit Service/Item Shop/coinService'; // Import coin service
-import { getTokenDetails } from '../../utils/tokenizer'; // Import the tokenizer
+import { createCustomer } from '../../service/customerService';
+import { createPaymentMethod } from '../../service/paymentMethodService';
+import { createSubscriptionPlan } from '../../service/recurringPaymentService';
 
 dotenv.config();
-
-const SUCCESS_RETURN_URL = 'https://example.com/payment-success';
-const FAILURE_RETURN_URL = 'https://example.com/payment-failure';
 
 // Function to fetch item details based on item ID
 export const getItemDetails = async (itemId: string) => {
@@ -31,51 +28,6 @@ const getUserDetailsByEmail = async (email: string) => {
   return user;
 };
 
-// Fetch coin balance
-export const getCoins = async (req: Request, res: Response) => {
-  const userId = parseInt(req.params.userId);
-
-  try {
-    const user = await User.findByPk(userId, {
-      attributes: ['coins'],
-    });
-
-    if (user) {
-      res.json({ coins: user.coins });
-    } else {
-      res.status(404).json({ error: 'User not found' });
-    }
-  } catch (error: any) {
-    res.status(500).json({ error: error.message });
-  }
-};
-
-// Deduct coins based on text input
-export const deductCoins = async (req: Request, res: Response) => {
-  const { userId, text } = req.body;
-
-  try {
-    // Calculate the number of tokens using the tokenizer
-    const { tokenCount, tokens } = getTokenDetails(text);
-
-    // Calculate the number of coins to deduct
-    const coinsToDeduct = tokenCount; // Assuming 1 coin per token
-
-    await deductCoinsByTokens(userId, text);
-    res.status(200).json({ 
-      message: 'Coins deducted successfully', 
-      coinsDeducted: coinsToDeduct, 
-      deductionDetails: `Deducted ${coinsToDeduct} coins for ${tokenCount} tokens (1 coin per token)`,
-      tokens: tokens 
-    });
-    console.log(`Input Text: ${text}`);
-    console.log(`Coins deducted: ${coinsToDeduct} coins for ${tokenCount} tokens`);
-    console.log(`Tokens: ${tokens.join(', ')}`);
-  } catch (error: any) {
-    res.status(500).json({ error: error.message });
-  }
-};
-
 export const buyItem = async (req: Request, res: Response) => {
   const { itemId, paymentMethod, email } = req.body;
 
@@ -92,33 +44,14 @@ export const buyItem = async (req: Request, res: Response) => {
     const date = new Date().toISOString().split('T')[0].replace(/-/g, ''); // Format date as YYYYMMDD
     orderId = `order-${itemId}-${item.name}-${user.username}-${date}`;
 
-    // Create payment method data
-    const paymentMethodData = {
-      type: 'ewallet',
-      reusability: 'ONE_TIME_USE',
-      ewallet: {
-        channel_code: paymentMethod,
-        channel_properties: {
-          success_return_url: SUCCESS_RETURN_URL,
-          failure_return_url: FAILURE_RETURN_URL,
-        },
-      },
-      customer_id: user.id.toString(), // Convert customer_id to string
-      metadata: {},
-      context: 'buy_item' as 'buy_item', // Explicitly set context as 'buy_item'
-    };
-
-    // Create payment method
-    const paymentMethodResponse = await createPaymentMethod(paymentMethodData);
-
     // Create a payment request
     const data: PaymentRequestParameters = {
       amount: item.price,
       paymentMethod: {
         ewallet: {
           channelProperties: {
-            successReturnUrl: SUCCESS_RETURN_URL,
-            failureReturnUrl: FAILURE_RETURN_URL,
+            successReturnUrl: 'https://example.com/payment-success',
+            failureReturnUrl: 'https://example.com/payment-failure',
           },
           channelCode: paymentMethod, // 'GCASH' or 'MAYA'
         },
@@ -166,5 +99,28 @@ Error: ${error.message}`);
     } else {
       res.status(500).json({ error: error.message });
     }
+  }
+};
+
+export const createSubscription = async (req: Request, res: Response) => {
+  try {
+    const { customerData, paymentMethodData, planData } = req.body;
+
+    // Create customer
+    const customer = await createCustomer(customerData);
+
+    // Create payment method
+    paymentMethodData.customer_id = customer.id;
+    const paymentMethod = await createPaymentMethod(paymentMethodData);
+
+    // Create subscription plan
+    planData.customer_id = customer.id;
+    planData.payment_methods = [{ payment_method_id: paymentMethod.id, rank: 1 }];
+    const subscriptionPlan = await createSubscriptionPlan(planData);
+
+    res.status(201).json(subscriptionPlan);
+  } catch (error) {
+    console.error('Error creating subscription:', error);
+    res.status(500).json({ message: 'Server error' });
   }
 };
