+++ backend/src/index.ts
@@ -10,11 +10,9 @@ import Jwt from 'jsonwebtoken';
 import routes from './routes/routes';
 import adminController from './routes/userCRUDRoutes';
 import * as authController from './controllers/authController';
-import invoiceRoutes from './routes/Xendit Routes/invoiceRoutes';
-import paymentRoutes from './routes/Xendit Routes/paymentRoutes';
+import invoiceRoutes from './routes/invoiceRoutes';
+import paymentRoutes from './routes/paymentRoutes';
 import User from './model/user';
-import shopRoutes from './routes/Xendit Routes/shopRoutes';
-import webhookRoutes from './routes/Xendit Routes/webhookRoutes';
 import chatRoutes from './routes/chatRoutes'; // Import the chatRoutes
 import shopRoutes from './routes/shopRoutes';
 import webhookRoutes from './routes/webhookRoutes';
