+++ backend/src/model/itemModel.ts
