+++ backend/src/model/user.ts
@@ -99,13 +99,10 @@ User.init(
             defaultValue: DataTypes.NOW,
         },
     },
-  },
-  {
-    sequelize,
-    modelName: 'User', // Optionally set the model name explicitly
-    tableName: 'Users', // Ensure the table name matches the database
-    timestamps: true, // Enable timestamps
-  }
+    {
+        sequelize,
+        modelName: 'User',
+    }
 );
 
 export default User;
