+++ backend/src/routes/shopRoutes.ts
@@ -1,9 +1,7 @@
 import express from 'express';
-import { createCustomer } from '../../service/Xendit Service/Subscription/customerService';
-import { createPaymentMethod } from '../../service/Xendit Service/Subscription/paymentMethodService';
-import { createSubscriptionPlan } from '../../service/Xendit Service/Subscription/subscriptionService';
-import { buyItem, getCoins, deductCoins } from '../../controllers/transaction_controllers/shopController';
-import { handlePaymentCallback } from '../../controllers/transaction_controllers/webhookController'; // Correct import path
+import { createCustomer } from '../service/customerService';
+import { createPaymentMethod } from '../service/paymentMethodService';
+import { createSubscriptionPlan } from '../service/subscriptionService';
 
 const router = express.Router();
 
@@ -55,11 +53,5 @@ router.post('/create-subscription-plan', async (req, res) => {
     res.status(400).json({ error: error.message });
   }
 });
-
-router.post('/buy-item', buyItem);
-router.post('/payment', handlePaymentCallback);
-router.get('/coins/:userId', getCoins);
-router.post('/deduct-coins', deductCoins);
-
 
 export default router;
